# openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Management API
  description: API for managing fleet vehicles, drivers, maintenance records, and assignments.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# No authentication required (optional; omit or add security schemes later)
security: []
servers:
  - url: http://localhost:5298/api
    description: Local API
  - url: /api
    description: Relative base path
...
components:
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable error description
...
    DriverStatus:
      type: integer
      enum: [0, 1, 2]
      description: 0 = Active, 1 = Inactive, 2 = Suspended
...
    AssignmentStatus:
      type: integer
      enum: [0, 1, 2]
      description: 0 = Active, 1 = Completed, 2 = Cancelled
...
    Driver:
      type: object
      description: Driver entity
      properties:
        id: { type: string, format: uuid, description: "Unique identifier (read-only)" }
        firstName: { type: string, description: Driver first name }
        lastName: { type: string, description: Driver last name }
        email: { type: string, nullable: true, description: Optional email }
        phoneNumber: { type: string, description: Contact phone number }
        licenseNumber: { type: string, description: Driver license number }
        licenseExpiryDate: { type: string, format: date-time, description: License expiry date }
        dateOfEmployment: { type: string, format: date-time, description: Employment start date }
        status: { $ref: '#/components/schemas/DriverStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDeleted: { type: boolean }
    DriverCreate:
      type: object
      description: Request body for creating a driver
      required: [firstName, lastName, phoneNumber, licenseNumber, licenseExpiryDate, dateOfEmployment, status]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, nullable: true }
        phoneNumber: { type: string }
        licenseNumber: { type: string }
        licenseExpiryDate: { type: string, format: date-time }
        dateOfEmployment: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/DriverStatus' }
...
    Assignment:
      type: object
      description: Assignment entity (vehicle-driver assignment)
      properties:
        id: { type: string, format: uuid, description: "Unique identifier (read-only)" }
        vehicleId: { type: string, format: uuid, description: ID of the assigned vehicle }
        driverId: { type: string, format: uuid, description: ID of the assigned driver }
        startDate: { type: string, format: date-time, description: Assignment start date }
        endDate: { type: string, format: date-time, description: Assignment end date }
        status: { $ref: '#/components/schemas/AssignmentStatus' }
        notes: { type: string, nullable: true, description: Optional notes }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDeleted: { type: boolean }
    AssignmentCreate:
      type: object
      description: Request body for creating an assignment
      required: [vehicleId, driverId, startDate, endDate, status]
      properties:
        vehicleId: { type: string, format: uuid }
        driverId: { type: string, format: uuid }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/AssignmentStatus' }
        notes: { type: string, nullable: true }
paths:
...
  /drivers:
    get:
      tags: [Drivers]
      summary: Get all drivers
      description: Returns all non-deleted drivers.
      operationId: getDrivers
      responses:
        '200':
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Driver' }
    post:
      tags: [Drivers]
      summary: Create driver
      description: Creates a new driver.
      operationId: createDriver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverCreate' }
      responses:
        '201':
          description: Driver created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Driver' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /drivers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Driver ID
    get:
      tags: [Drivers]
      summary: Get driver by ID
      description: Returns a single driver by ID.
      operationId: getDriver
      responses:
        '200':
          description: Driver found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Driver' }
        '404':
          description: Driver not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Drivers]
      summary: Update driver
      description: Updates an existing driver. Body id must match path id.
      operationId: updateDriver
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Driver' }
      responses:
        '204':
          description: Update successful
        '400':
          description: Driver ID mismatch
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Driver not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Drivers]
      summary: Delete driver
      description: Soft-deletes the driver.
      operationId: deleteDriver
      responses:
        '204':
          description: Soft-delete successful
        '404':
          description: Driver not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /drivers/{id}/assignments:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Driver ID
    get:
      tags: [Drivers]
      summary: Get driver assignments
      description: Returns all non-deleted assignments for the given driver.
      operationId: getDriverAssignments
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
        '404':
          description: Driver not found or no assignments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /assignments:
    get:
      tags: [Assignments]
      summary: Get all assignments
      description: Returns all non-deleted assignments.
      operationId: getAssignments
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
    post:
      tags: [Assignments]
      summary: Create assignment
      description: Creates a new assignment. Vehicle and driver must exist and must not already have an active assignment. Start date cannot be in the past; end date must be after start date.
      operationId: createAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentCreate' }
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '400':
          description: Vehicle/driver not found or validation errors (e.g. dates, duplicate active assignment)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /assignments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Assignment ID
    get:
      tags: [Assignments]
      summary: Get assignment by ID
      description: Returns a single assignment by ID.
      operationId: getAssignment
      responses:
        '200':
          description: Assignment found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Assignments]
      summary: Update assignment
      description: Updates an existing assignment. Same business rules as create for vehicle/driver and active-assignment uniqueness.
      operationId: updateAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Assignment' }
      responses:
        '204':
          description: Update successful
        '400':
          description: ID mismatch or validation errors
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Assignments]
      summary: Delete assignment
      description: Soft-deletes the assignment.
      operationId: deleteAssignment
      responses:
        '204':
          description: Soft-delete successful
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /assignments/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Assignment ID
    patch:
      tags: [Assignments]
      summary: Complete assignment
      description: Sets assignment status to Completed and end date to current UTC time.
      operationId: completeAssignment
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /assignments/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Assignment ID
    patch:
      tags: [Assignments]
      summary: Cancel assignment
      description: Sets assignment status to Cancelled.
      operationId: cancelAssignment
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }